requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
google-genai>=0.3.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
httpx>=0.25.0
python-multipart>=0.0.6
